+++ apps/ungdomsytelse-deltaker/src/utils/initApiClients.ts
@@ -3,20 +3,23 @@ import { EnvKey, getRequiredEnv } from '@navikt/sif-common-env';
 import { initUngBrukerdialogApiClient } from '@navikt/ung-brukerdialog-api';
 import { initUngDeltakelseOpplyserApiDeltakerClient } from '@navikt/ung-deltakelse-opplyser-api-deltaker';
 
+import { setRedirectingToLogin } from '../sentry/instrument';
+
 export const initApiClients = () => {
     initUngDeltakelseOpplyserApiDeltakerClient({
-        onUnAuthorized: () => {
-            globalThis.location.reload();
-        },
+        frontendPath: getRequiredEnv(EnvKey.UNG_DELTAKELSE_OPPLYSER_FRONTEND_PATH),
+        loginURL: getRequiredEnv(EnvKey.SIF_PUBLIC_LOGIN_URL),
+        onUnauthorized: setRedirectingToLogin,
     });
 
     initUngBrukerdialogApiClient({
-        onUnAuthorized: () => {
-            globalThis.location.reload();
-        },
+        frontendPath: getRequiredEnv(EnvKey.UNG_BRUKERDIALOG_API_FRONTEND_PATH),
+        loginURL: getRequiredEnv(EnvKey.SIF_PUBLIC_LOGIN_URL),
+        onUnauthorized: setRedirectingToLogin,
     });
     initK9BrukerdialogProsesseringApiClients({
         frontendPath: getRequiredEnv(EnvKey.K9_BRUKERDIALOG_PROSESSERING_FRONTEND_PATH),
-        loginURL: '#',
+        loginURL: getRequiredEnv(EnvKey.SIF_PUBLIC_LOGIN_URL),
+        onUnauthorized: setRedirectingToLogin,
     });
 };