+++ apps/aktivitetspenger-innsyn/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/aktivitetspenger-innsyn
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-sentry@0.28.65
+    - @navikt/sif-common-core-ds@10.3.55
+    - @sif/ung-innsyn@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes